Fixed comments from Pull request: V2 being committed
